Table component refactoring
