+++ src/ui/table/constants/constants.ts
@@ -1,0 +1,3 @@
+export const PAGINATION_QUERY_PARAM_NAME = "page";
+
+export const ROWS_PER_PAGE = 10;
